Refactor project structure by renaming and moving files; remove unused components and services; add react-redux dependency
