+++ client/package.json
@@ -24,6 +24,7 @@
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
     "react-icons": "^5.3.0",
+    "react-redux": "^9.1.2",
     "react-router-dom": "^6.28.0",
     "redux": "^5.0.1",
     "tailwind-merge": "^2.5.5",
