--- client/src/components/page/NetworkError.tsx
@@ -1,14 +1,0 @@
-export default function NetworkError () {
-    return (
-        <div className="w-full h-screen flex items-center justify-center">
-            <div>
-                <div>
-                    Problemme de connexion.
-                </div>
-                <div>
-                    Veuillez vérifier votre connexion internet.
-                </div>
-            </div>
-        </div>
-    )
-}
