--- client/src/components/pwa/PWABadge.css
@@ -1,29 +1,0 @@
-.PWABadge-container {
-    padding: 0;
-    margin: 0;
-    width: 0;
-    height: 0;
-}
-.PWABadge-toast {
-    position: fixed;
-    right: 0;
-    bottom: 0;
-    margin: 16px;
-    padding: 12px;
-    border: 1px solid #8885;
-    border-radius: 4px;
-    z-index: 1;
-    text-align: left;
-    box-shadow: 3px 4px 5px 0 #8885;
-    background-color: white;
-}
-.PWABadge-toast-message {
-    margin-bottom: 8px;
-}
-.PWABadge-toast-button {
-    border: 1px solid #8885;
-    outline: none;
-    margin-right: 5px;
-    border-radius: 2px;
-    padding: 3px 10px;
-}
