+++ client/src/main.tsx
@@ -3,14 +3,11 @@ import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
 import { BrowserRouter } from 'react-router-dom'
-import { NetworkProvider } from './core/contexts/NetworkContext.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
      <BrowserRouter>
-      <NetworkProvider>
         <App />
-      </NetworkProvider>
      </BrowserRouter>
   </React.StrictMode>,
 )
