--- client/src/modules/auth/login/pages/Login.tsx
@@ -1,19 +1,0 @@
-import LoginForm from "../components/LoginForm";
-
-
-export default function Login() {
-    return (
-        <>
-            <div>
-                {/* En-tête */}
-
-                {/* Contenu */}
-                <main>
-                    <LoginForm />
-                </main>
-
-                {/* Pied de page */}
-            </div>
-        </>
-    )
-}
