--- client/src/modules/auth/register/pages/Register.tsx
@@ -1,31 +1,0 @@
-import RegisterForm from "../components/RegisterForm";
-
-export default function Register() {
-    /**
-     * ! STATE (état, données) de l'application
-     */
-
-
-    /**
-     * ! COMPORTEMENT (méthodes, fonctions) de l'application
-     */
-
-
-    /**
-     * ! AFFICHAGE (render) de l'application
-     */
-    return (
-        <>
-            <div>
-                {/* En-tête */}
-
-                {/* Contenu */}
-                <main>
-                    <RegisterForm />
-                </main>
-
-                {/* Pied de page */}
-            </div>
-        </>
-    )
-}
