+++ client/src/pages/app/day_photos/DayPhotos.tsx
