+++ client/src/pages/app/days/components/ImageDetail.tsx
