+++ client/src/pages/app/error/NetworkErrorDialog.tsx
